+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for passing a sequence of [hikari.CommandChoice][hikari.commands.CommandChoice]s
+  to `choices` when adding an option to a slash command.
+
+### Changed
+- Bumped the minimum Hikari version to `2.0.0.dev118`.
+
 ## [2.12.0] - 2023-03-12
 ### Added
 - Support for declaring the options for a command using